[v0.89.1][maintainability] Split cli/run_artifacts/runtime.rs by responsibility (Closes #1945)
