+++ adl/src/cli/run_artifacts/runtime/resume_state.rs
@@ -1,0 +1,169 @@
+use super::super::*;
+
+pub(crate) fn load_resume_state(
+    path: &Path,
+    resolved: &resolve::AdlResolved,
+) -> Result<execute::ResumeState> {
+    let raw = std::fs::read_to_string(path)
+        .with_context(|| format!("failed to read resume state '{}'", path.display()))?;
+    let artifact: RunStateArtifact = serde_json::from_str(&raw).with_context(|| {
+        format!(
+            "failed to parse resume state '{}' as run_state artifact",
+            path.display()
+        )
+    })?;
+
+    if artifact.schema_version != RUN_STATE_SCHEMA_VERSION {
+        return Err(anyhow::anyhow!(
+            "resume state schema_version mismatch: state='{}' expected='{}'",
+            artifact.schema_version,
+            RUN_STATE_SCHEMA_VERSION
+        ));
+    }
+
+    if artifact.schema_version != RUN_STATE_SCHEMA_VERSION {
+        return Err(anyhow::anyhow!(
+            "resume state schema_version mismatch: state='{}' expected='{}'",
+            artifact.schema_version,
+            RUN_STATE_SCHEMA_VERSION
+        ));
+    }
+
+    if artifact.status != "paused" {
+        return Err(anyhow::anyhow!(
+            "resume state must have status='paused' (found='{}' for run_id='{}' in '{}')",
+            artifact.status,
+            artifact.run_id,
+            path.display()
+        ));
+    }
+    if artifact.run_id != resolved.run_id {
+        return Err(anyhow::anyhow!(
+            "resume run_id mismatch: state='{}' current='{}'",
+            artifact.run_id,
+            resolved.run_id
+        ));
+    }
+    if artifact.workflow_id != resolved.workflow_id {
+        return Err(anyhow::anyhow!(
+            "resume workflow_id mismatch for run_id='{}' in '{}': state='{}' current='{}'",
+            artifact.run_id,
+            path.display(),
+            artifact.workflow_id,
+            resolved.workflow_id
+        ));
+    }
+    if artifact.version != resolved.doc.version {
+        return Err(anyhow::anyhow!(
+            "resume version mismatch for run_id='{}' in '{}': state='{}' current='{}'",
+            artifact.run_id,
+            path.display(),
+            artifact.version,
+            resolved.doc.version
+        ));
+    }
+    let plan_hash = execution_plan_hash(&resolved.execution_plan)?;
+    if artifact.execution_plan_hash != plan_hash {
+        return Err(anyhow::anyhow!(
+            "resume execution plan mismatch for run_id='{}' in '{}'; state plan != current plan (resume requires identical plan + ordering)",
+            artifact.run_id,
+            path.display()
+        ));
+    }
+    let pause = artifact
+        .pause
+        .ok_or_else(|| anyhow::anyhow!("resume state missing pause payload"))?;
+
+    let completed_step_ids = pause.completed_step_ids.into_iter().collect();
+    Ok(execute::ResumeState {
+        completed_step_ids,
+        saved_state: pause.saved_state,
+        completed_outputs: pause.completed_outputs,
+        steering_history: artifact.steering_history,
+    })
+}
+
+pub(crate) fn resume_state_path_for_run_id(run_id: &str) -> Result<PathBuf> {
+    Ok(artifacts::RunArtifactPaths::for_run(run_id)?.pause_state_json())
+}
+
+pub(crate) fn load_pause_state_artifact(path: &Path) -> Result<PauseStateArtifact> {
+    let raw = std::fs::read_to_string(path)
+        .with_context(|| format!("failed to read pause state '{}'", path.display()))?;
+    let artifact: PauseStateArtifact =
+        serde_json::from_str(&raw).with_context(|| "failed to parse pause_state.json")?;
+    Ok(artifact)
+}
+
+pub(crate) fn load_steering_patch(path: &Path) -> Result<(execute::SteeringPatch, String)> {
+    let raw = std::fs::read(path)
+        .with_context(|| format!("failed to read steering patch '{}'", path.display()))?;
+    let fingerprint = stable_fingerprint_hex(&raw);
+    let patch: execute::SteeringPatch =
+        serde_json::from_slice(&raw).with_context(|| "failed to parse steering patch JSON")?;
+    execute::validate_steering_patch(&patch)?;
+    Ok((patch, fingerprint))
+}
+
+pub(crate) fn validate_pause_artifact_basic(
+    artifact: &PauseStateArtifact,
+    run_id: &str,
+) -> Result<()> {
+    if artifact.schema_version != PAUSE_STATE_SCHEMA_VERSION {
+        return Err(anyhow::anyhow!(
+            "pause state schema_version mismatch: state='{}' expected='{}'",
+            artifact.schema_version,
+            PAUSE_STATE_SCHEMA_VERSION
+        ));
+    }
+    if artifact.status != "paused" {
+        return Err(anyhow::anyhow!(
+            "pause state must have status='paused' (found '{}')",
+            artifact.status
+        ));
+    }
+    if artifact.run_id != run_id {
+        return Err(anyhow::anyhow!(
+            "pause state run_id mismatch: state='{}' requested='{}'",
+            artifact.run_id,
+            run_id
+        ));
+    }
+    Ok(())
+}
+
+pub(crate) fn validate_pause_artifact_for_resume(
+    artifact: &PauseStateArtifact,
+    run_id: &str,
+    resolved: &resolve::AdlResolved,
+) -> Result<()> {
+    validate_pause_artifact_basic(artifact, run_id)?;
+    if artifact.run_id != resolved.run_id {
+        return Err(anyhow::anyhow!(
+            "resume run_id mismatch: state='{}' current='{}'",
+            artifact.run_id,
+            resolved.run_id
+        ));
+    }
+    if artifact.workflow_id != resolved.workflow_id {
+        return Err(anyhow::anyhow!(
+            "resume workflow_id mismatch: state='{}' current='{}'",
+            artifact.workflow_id,
+            resolved.workflow_id
+        ));
+    }
+    if artifact.version != resolved.doc.version {
+        return Err(anyhow::anyhow!(
+            "resume version mismatch: state='{}' current='{}'",
+            artifact.version,
+            resolved.doc.version
+        ));
+    }
+    let plan_hash = execution_plan_hash(&resolved.execution_plan)?;
+    if artifact.execution_plan_hash != plan_hash {
+        return Err(anyhow::anyhow!(
+            "resume execution plan hash mismatch; resume requires identical plan and ordering"
+        ));
+    }
+    Ok(())
+}
